Working on render pass and changed initialization

+++ src/engine/graphics/Descriptor.cpp
@@ -72,7 +72,7 @@ namespace Atlas {
             VkDescriptorPoolCreateInfo poolInfo = {};
             poolInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
             poolInfo.flags = 0;
-            poolInfo.maxSets = sizes.size() * DESCRIPTOR_POOL_SIZE;
+            poolInfo.maxSets = uint32_t(sizes.size()) * DESCRIPTOR_POOL_SIZE;
             poolInfo.poolSizeCount = uint32_t(sizes.size());
             poolInfo.pPoolSizes = sizes.data();
